♻️ Refactor: 번개 목록의 컴포넌트 분리
